refactor : remove unused directory
