--- src/features/review/components/index.js
@@ -1,3 +1,0 @@
-export { default as ReviewInput } from './ReviewInput';
-export { default as ReviewScrollView } from './ReviewScrollView';
-export { default as ReviewStarRatingSelector } from './ReviewStarRatingSelector';
