+++ src/features/review/index.js
@@ -1,1 +1,3 @@
-export * from './components';
+export { default as ReviewInput } from './ReviewInput';
+export { default as ReviewScrollView } from './ReviewScrollView';
+export { default as ReviewStarRatingSelector } from './ReviewStarRatingSelector';
